+++ NEWS.md
@@ -1,5 +1,8 @@
 # cpsR (development version)
 
+* `get_asec()` now supports CPS ASEC microdata for 2022.
+* `get_basic()` now supports basic monthly CPS microdata for 1989 to 1993.
+
 # cpsR 0.5.0
 
 * `get_basic()` now supports basic monthly CPS microdata for 2022.